mmr.go: update and test Random sampling distribution

+++ gommr/mmr.go
@@ -504,7 +504,7 @@ func (m *mmr) genProof3(right_difficulty *big.Int) *Proof3 {
 
 	weights, blocks := []float64{}, []uint64{}
 	for i := 0; i < int(required_queries); i++ {
-		h := RlpHash([]interface{}{root_hash, i})
+		h := RlpHash([]interface{}{root_hash, uint64(i)})
 		random := Hash_to_f64(h)
 		r3, _ := new(big.Float).SetInt(m.getRootDifficulty()).Float64()
 		aggr_weight := cdf(random, vd_calculate_delta(r1, r3))
@@ -801,7 +801,7 @@ func (m *mmr) CreateNewProof(right_difficulty *big.Int) (*ProofInfo, []uint64, [
 
 	weights, blocks := []float64{}, []uint64{}
 	for i := 0; i < int(required_queries); i++ {
-		h := RlpHash([]interface{}{root_hash, i})
+		h := RlpHash([]interface{}{root_hash, uint64(i)})
 		random := Hash_to_f64(h)
 		r3, _ := new(big.Float).SetInt(m.getRootDifficulty()).Float64()
 		aggr_weight := cdf(random, vd_calculate_delta(r1, r3))
@@ -995,7 +995,7 @@ func verify_required_blocks(blocks []uint64, root_hash Hash, root_difficulty, ri
 	}
 	weights := []float64{}
 	for i := 0; i < int(required_queries); i++ {
-		h := RlpHash([]interface{}{root_hash, i})
+		h := RlpHash([]interface{}{root_hash, uint64(i)})
 		random := Hash_to_f64(h)
 		r3, _ := new(big.Float).SetInt(root_difficulty).Float64()
 		aggr_weight := cdf(random, vd_calculate_delta(r1, r3))
